iutctl: Unify iutctl for zephyr and mynewt
Unify ZephyrCtl and MynewtCtl to make it easier to add and maintain
new features.
