+++ autopts/ptsprojects/zephyr/ztestcase.py
@@ -31,21 +31,12 @@ class ZTestCase(TestCaseLT1):
         self.stack = get_stack()
         self.zephyrctl = get_iut()
 
+        # Init stack.core to be able to receive IUT ready event
         self.cmds.insert(0, TestFunc(self.stack.core_init))
-
-        # For HW, the IUT ready event is triggered at its reset and
-        # can be only received after socket is successfully running.
-        # This hw_reset will be active only at the first test case.
-        self.cmds.insert(1, TestFunc(self.zephyrctl.hw_reset))
-
-        # This will open BTP socket and start QEMU process.
-        # For QEMU, the IUT ready event is sent at startup of the process.
-        self.cmds.insert(2, TestFunc(self.zephyrctl.start, self))
-
-        # Now socket should be open for IUT ready event from HW.
-        # This hw_reset will be active only at the first test case.
-        self.cmds.insert(3, TestFunc(self.zephyrctl.hw_reset))
-        self.cmds.insert(4, TestFunc(self.zephyrctl.wait_iut_ready_event, False))
+        # Open BTP socket and start IUT
+        self.cmds.insert(1, TestFunc(self.zephyrctl.start, self))
+        # Await IUT ready event
+        self.cmds.insert(2, TestFunc(self.zephyrctl.wait_iut_ready_event, False))
 
         self.cmds.append(TestFuncCleanUp(self.stack.cleanup))
